use generated RELEASE_FLAVOR
Change of variable name from FLAVOR to RELEASE_FLAVOR for many builds.
Consequently many job files needed to include a dependency.

        modified:   transfer/export_directories_dpkg.sh
        modified:   transfer/export_docbook_branch.sh
        modified:   transfer/export_docbook_tag.sh
        modified:   transfer/export_package_debian.sh
        modified:   transfer/export_python_branch_srpm_scientific.sh
        modified:   transfer/export_python_tag_deb_stdeb.sh

+++ transfer/export_package_debian.sh
@@ -3,8 +3,20 @@ if [ "X${DIR_EXPORT_ROOT}" = "X" ] ; then
     sleep 10
     exit 1
 fi
+if [ "X${RELEASE_FLAVOR}" = "X" ] ; then
+    echo "RELEASE_FLAVOR not defined"
+    sleep 10
+    exit 1
+fi
+
+
+
 tar -zxvf artifacts.tgz
 rm -f artifacts.tgz
+
+
+
+
 RELEASE_TYPE="nightly"
 #if "X$RELEASE" == "Xdevelopment" then
 #    RELEASE_TYPE="nightly"
@@ -14,10 +26,9 @@ if [ "X$RELEASE" = "Xproduction" ] ; then
 fi
 
 PLATFORM="x86_64"
-FLAVOR="debian/7"
 mkdir -p ${DIR_EXPORT_ROOT}
 dir_tgz="${DIR_EXPORT_ROOT}/${FLAVOR}/${RELEASE_TYPE}/src/tgz"
-dir_btgz="${DIR_EXPORT_ROOT}/${FLAVOR}/${RELEASE_TYPE}/${PLATFORM}/tgz"
+dir_btgz="${DIR_EXPORT_ROOT}/${RELEASE_FLAVOR}/${RELEASE_TYPE}/${PLATFORM}/tgz"
 mkdir -p ${dir_tgz}
 mkdir -p ${dir_btgz}
 /usr/bin/rsync -v --ignore-existing build/dist/*.src.tar.gz \
