+++ transfer/export_python_tag_deb_stdeb.sh
@@ -3,6 +3,11 @@ if [ "X${DIR_EXPORT_ROOT}" = "X" ] ; then
     sleep 10
     exit 1
 fi
+if [ "X${RELEASE_FLAVOR}" = "X" ] ; then
+    echo "RELEASE_FLAVOR not defined"
+    sleep 10
+    exit 1
+fi
 tar -zxvf artifacts.tgz
 rm -f artifacts.tgz
 RELEASE_TYPE="nightly"
@@ -14,11 +19,10 @@ if [ "X$RELEASE" = "Xproduction" ] ; then
 fi
 
 PLATFORM="x86_64"
-FLAVOR="debian/7"
 mkdir -p ${DIR_EXPORT_ROOT}
-dir_tgz="${DIR_EXPORT_ROOT}/${FLAVOR}/${RELEASE_TYPE}/src/tgz"
-dir_btgz="${DIR_EXPORT_ROOT}/${FLAVOR}/${RELEASE_TYPE}/${PLATFORM}/tgz"
-dir_deb="${DIR_EXPORT_ROOT}/${FLAVOR}/${RELEASE_TYPE}/${PLATFORM}/deb"
+dir_tgz="${DIR_EXPORT_ROOT}/${RELEASE_FLAVOR}/${RELEASE_TYPE}/src/tgz"
+dir_btgz="${DIR_EXPORT_ROOT}/${RELEASE_FLAVOR}/${RELEASE_TYPE}/${PLATFORM}/tgz"
+dir_deb="${DIR_EXPORT_ROOT}/${RELEASE_FLAVOR}/${RELEASE_TYPE}/${PLATFORM}/deb"
 mkdir -p ${dir_tgz}
 mkdir -p ${dir_btgz}
 mkdir -p ${dir_deb}
